SocialList.tsx to ~/content/config.tsx
