+++ src/content/config.tsx
@@ -1,6 +1,9 @@
+import { defineCollection, z } from "astro:content";
+
 //import { h } from 'preact'
-import { TwitchIcon, YoutubeIcon, FacebookIcon, LucideIcon, LucideProps } from 'lucide-preact'
-import DiscordIcon from "./DiscordIcon"
+import { TwitchIcon, YoutubeIcon, FacebookIcon } from 'lucide-preact'
+import type { LucideIcon, LucideProps } from 'lucide-preact'
+import DiscordIcon from "../components/DiscordIcon"
 
 export interface social {
   link: string
@@ -53,3 +56,17 @@ export const socials: social[] = [
     right: "-right-[70px]"
   },
 ]
+
+const showcase = defineCollection({
+  type: "data",
+  schema: z.object({
+    title: z.string().min(1),
+    image: z.string(),
+    url: z.string().url(),
+    featured: z.number().min(1).optional(),
+  }),
+});
+
+export const collections = {
+  showcase,
+};
